backend/requirements.txt
boto3
fastapi
uvicorn
strands-agents
strands-agents-tools
pydantic
